feat: (pool): Add auto earnings interest breakdown to pool cards
