+++ src/views/Pools/components/CakeVaultCard/RecentCakeProfitRow.tsx
@@ -3,19 +3,19 @@ import { useWeb3React } from '@web3-react/core'
 import { useTranslation } from 'contexts/Localization'
 import { usePriceCakeBusd } from 'state/farms/hooks'
 import { useVaultPoolByKey } from 'state/pools/hooks'
-import { VaultKey } from 'state/types'
+import { DeserializedPool } from 'state/types'
 import { getCakeVaultEarnings } from 'views/Pools/helpers'
 import RecentCakeProfitBalance from './RecentCakeProfitBalance'
 
-const RecentCakeProfitCountdownRow = ({ vaultKey }: { vaultKey: VaultKey }) => {
+const RecentCakeProfitCountdownRow = ({ pool }: { pool: DeserializedPool }) => {
   const { t } = useTranslation()
   const { account } = useWeb3React()
   const {
     pricePerFullShare,
-    userData: { cakeAtLastUserAction, userShares, lastUserActionTime },
-  } = useVaultPoolByKey(vaultKey)
+    userData: { cakeAtLastUserAction, userShares },
+  } = useVaultPoolByKey(pool.vaultKey)
   const cakePriceBusd = usePriceCakeBusd()
-  const { hasAutoEarnings, autoCakeToDisplay, autoUsdToDisplay } = getCakeVaultEarnings(
+  const { hasAutoEarnings, autoCakeToDisplay } = getCakeVaultEarnings(
     account,
     cakeAtLastUserAction,
     userShares,
@@ -23,20 +23,10 @@ const RecentCakeProfitCountdownRow = ({ vaultKey }: { vaultKey: VaultKey }) => {
     cakePriceBusd.toNumber(),
   )
 
-  const lastActionInMs = lastUserActionTime && parseInt(lastUserActionTime) * 1000
-  const dateTimeLastAction = new Date(lastActionInMs)
-  const dateStringToDisplay = dateTimeLastAction.toLocaleString()
-
   return (
     <Flex alignItems="center" justifyContent="space-between">
       <Text fontSize="14px">{`${t('Recent CAKE profit')}:`}</Text>
-      {hasAutoEarnings && (
-        <RecentCakeProfitBalance
-          cakeToDisplay={autoCakeToDisplay}
-          dollarValueToDisplay={autoUsdToDisplay}
-          dateStringToDisplay={dateStringToDisplay}
-        />
-      )}
+      {hasAutoEarnings && <RecentCakeProfitBalance cakeToDisplay={autoCakeToDisplay} pool={pool} account={account} />}
     </Flex>
   )
 }
